Close PreparedStatement and RestultSet object where possible.
In many cases these were already being closed.  Where that was not true, this pass mostly just takes care of the "low-hanging fruit" -- but that should eliminate most of the close activity from the finalizer thread.  Added one javadoc to make clear that the caller was responsible for cleanup.

A couple TODO comments were added, and one expanded.